backend/requirements.txt
yt-dlp
fastapi
uvicorn
python-multipart
spotipy
python-dotenv
requests
